Explicit 64->32 bit conversion in repo_cache.cpp
On Fedora 41 i686, this conversion causes a warning which prevents
compiling, so we need to make it explicit with static_cast. But it
represents a real bug: get_bytes_removed() can overflow on 32-bit
systems. We should use a 64-bit integer for get_bytes_removed(), but
changing this would break ABI compatibility since the just-released DNF
5.2.13.0.

+++ libdnf5/repo/repo_cache.cpp
@@ -50,7 +50,8 @@ std::size_t remove(
             size = 0;
         }
         if (std::filesystem::remove(path)) {
-            bytes_count += size;
+            // Note: bytes_count will overflow on 32-bit systems if the total space removed exceeds 4GiB.
+            bytes_count += static_cast<std::size_t>(size);
             return 1;
         }
     } catch (const std::filesystem::filesystem_error & ex) {
